Stupid compilers don't like when we mix uint16_t:14 and uint8_t:1 in the same struct

+++ trunk/src/open9xeeprom.h
@@ -238,9 +238,9 @@ PACK(typedef struct t_Open9xPhaseData_v201 {
 
 PACK(typedef struct t_Open9xTimerData_v201 {
   int8_t    mode;            // timer trigger source -> off, abs, stk, stk%, sw/!sw, !m_sw/!m_sw
-  uint16_t  val:14;
-  uint8_t   persistent:1;
-  uint8_t   dir:1;             // 0=>Count Down, 1=>Count Up
+  uint16_t   val:14;
+  uint16_t   persistent:1;
+  uint16_t   dir:1;          // 0=>Count Down, 1=>Count Up
 
   operator TimerData();
   t_Open9xTimerData_v201() { memset(this, 0, sizeof(t_Open9xTimerData_v201)); }
